+++ src/test/java/de/csbdresden/csbdeep/network/InputMapperTest.java
@@ -30,7 +30,8 @@ public class InputMapperTest {
 	public void network2D2D_input2D() {
 		ImageJ ij = new ImageJ();
 		Dataset input = ij.dataset().create(new FloatType(), new long[]{2,3}, "", new AxisType[]{Axes.X, Axes.Y});
-		Network network = new TestNetworkSingle2D2D(ij.get(TensorFlowService.class), ij.dataset());
+		Network network = new TestNetworkSingle2D2D();
+		ij.context().inject(network);
 
 		doMapping(input, network);
 
@@ -53,7 +54,8 @@ public class InputMapperTest {
 	public void network2D2D_input3D() {
 		ImageJ ij = new ImageJ();
 		Dataset input = ij.dataset().create(new FloatType(), new long[]{2,3,4}, "", new AxisType[]{Axes.X, Axes.Y, Axes.Z});
-		Network network = new TestNetworkSingle2D2D(ij.get(TensorFlowService.class), ij.dataset());
+		Network network = new TestNetworkSingle2D2D();
+		ij.context().inject(network);
 
 		doMapping(input, network);
 
@@ -77,7 +79,8 @@ public class InputMapperTest {
 	public void network3D2D_input3D() {
 		ImageJ ij = new ImageJ();
 		Dataset input = ij.dataset().create(new FloatType(), new long[]{2,3,4}, "", new AxisType[]{Axes.X, Axes.Y, Axes.Z});
-		Network network = new TestNetworkSingle3D2D(ij.get(TensorFlowService.class), ij.dataset());
+		Network network = new TestNetworkSingle3D2D();
+		ij.context().inject(network);
 
 		doMapping(input, network);
 
@@ -102,7 +105,8 @@ public class InputMapperTest {
 	public void single3D3DTest() {
 		ImageJ ij = new ImageJ();
 		Dataset input = ij.dataset().create(new FloatType(), new long[]{2,3,4}, "", new AxisType[]{Axes.X, Axes.Y, Axes.Z});
-		Network network = new TestNetworkSingle3D3D(ij.get(TensorFlowService.class), ij.dataset());
+		Network network = new TestNetworkSingle3D3D();
+		ij.context().inject(network);
 
 		doMapping(input, network);