+++ src/test/java/de/csbdresden/csbdeep/network/TestNetwork.java
@@ -1,13 +1,10 @@
 
 package de.csbdresden.csbdeep.network;
 
-import org.tensorflow.framework.TensorInfo;
-
 import de.csbdresden.csbdeep.network.model.tensorflow.TensorFlowNetwork;
 import net.imagej.Dataset;
-import net.imagej.DatasetService;
-import net.imagej.tensorflow.TensorFlowService;
 import net.imglib2.type.numeric.RealType;
+import org.tensorflow.framework.TensorInfo;
 
 public abstract class TestNetwork<T extends RealType<T>> extends
 		TensorFlowNetwork<T>
@@ -18,10 +15,9 @@ public abstract class TestNetwork<T extends RealType<T>> extends
 	protected int inputCount = 1;
 	protected int outputCount = 1;
 
-	public TestNetwork(TensorFlowService tensorFlowService,
-	                   DatasetService datasetService)
+	public TestNetwork()
 	{
-		super(tensorFlowService, datasetService, null);
+		super(null);
 	}
 
 	@Override