+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 # Changelog
 
-### Version 0.3.2
+## 0.5.0
+- updated to `imagej-tensorflow 1.1.4`
+  - native JNIs can now be switched via `Edit > Options > TensorFlow...`
+
+## 0.4.1
+-  `GenericCoreNetwork`: remove menu path
+
+## 0.4.0
+- Adds `GenericTableNetwork`
+  - add `scijava-plugins-io-table` dependency
+  - rename `DefaultOutputProcessor` to `DatasetOutputProcessor`
+  - add `TableOutputProcessor`
+  - move everything but the `output` Parameter and the initialization of the `OutputProcessor` from `GenericNetwork` to `GenericCoreNetwork`
+  - `GenericCoreNetwork`: `protected abstract initOutputProcessor` method
+  - `GenericCoreNetwork`: `protected abstract computeOutput` method
+  - make `GenericNetwork` extending `GenericCoreNetwork` with `Dataset` output parameter
+  - create `GenericTableNetwork` extending `GenericCoreNetwork` with `GenericTable` output parameter
+  - `OutputProcessor` is typed now, changed signature of `run`
+  - `ImageTensor`: add `tilingAllowed` attribute
+  - `ImageTensor`: make a few methods private
+- `POM`: `pom-scijava 23.2.0` &rarr; `27.0.1`
+-  `PercentileNormalizer`: Fix normalization for images with the same value for all pixels
+  - Images with the same value were normalized to images of value `NaN`
+  - now they return an image of value `0.0`
+
+## 0.3.4
+- don't load `meta.json` of model since we do not use it yet
+
+## 0.3.2
 - !!! `groupId` and `artifactId` changed !!!
 ```xml
 <groupId>de.csbdresden</groupId>